backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
bcrypt==4.1.3
PyJWT==2.8.0
supabase==2.4.6
